+++ src/hello/components/Hello.test.js
@@ -1,9 +1,7 @@
-// @flow
-
 import React from 'react'
 import renderer from 'react-test-renderer'
 import {Hello} from './Hello'
-import {sayHello} from './actions'
+import {sayHello} from '../actions'
 
 test('Hello component should set message from props', () => {
   const tree = renderer.create(
@@ -11,13 +9,3 @@ test('Hello component should set message from props', () => {
   ).toJSON()
   expect(tree).toMatchSnapshot()
 })
-
-function typeTests(){
-  // $FlowExpectError
-  const a = <Hello />
-
-  // $FlowExpectError
-  const b = <Hello message="Hello from Jest." />
-
-  const c = <Hello sayHello={sayHello} message="Hello from Jest." />
-}
